+++ module/USR.py
@@ -15,15 +15,7 @@ class USR(Operator):
         if not os.path.exists(output):
             os.makedirs(output)
         w, h, fps = self.extractParameters(input)
-        
-        # convert yuv to png
-        tmp_folder = os.path.join(output, 'tmp')
-        if not os.path.exists(tmp_folder):
-            os.mkdir(tmp_folder)
-        cmd = 'ffmpeg -framerate {} -s {}x{} -i {} -f image2 {}/%d.png -hide_banner'.format(
-            fps, w, h, input, tmp_folder)
-        os.system(cmd)
-        # use USR to super resolutioin
+        # use get up rate
         obj = re.match(r'.*/downResolutionStage_FFmpeg_downRate_(.+)/enc.*', input)
         if obj:
             para = obj.group(1)
@@ -31,17 +23,31 @@ class USR(Operator):
         else:
             raise Exception('USR intput valid:{}'.format(input))
 
-        cmd = 'cd {}/3rdparty/VideoPhotoRepair;python main.py {} --operation usr --scale {} --clc'.format(
-            os.path.dirname(__file__), os.path.abspath(tmp_folder), int(up_rate) )
-      
-        os.system(cmd)
-        # convert png to yuv
-        newFileName = self.generateFileName(int(w*up_rate), int(h*up_rate), fps)
-        output = os.path.join(output, newFileName)
-        cmd = 'ffmpeg -i {}/3rdparty/VideoPhotoRepair/results/usr/%d.png -pix_fmt yuv420p {} -hide_banner'.format(
-            os.path.dirname(__file__), output)
-        os.system(cmd)       
-        shutil.rmtree(tmp_folder)
+        if abs(up_rate - 1) < 1e-5:
+            newFileName = self.generateFileName(int(w*up_rate), int(h*up_rate), fps)
+            output = os.path.join(output, newFileName)
+            os.symlink(os.path.abspath(input), output)
+        else:
+            # convert yuv to png
+            tmp_folder = os.path.join(output, 'tmp')
+            if not os.path.exists(tmp_folder):
+                os.mkdir(tmp_folder)
+            cmd = 'ffmpeg -framerate {} -s {}x{} -i {} -f image2 {}/%d.png -hide_banner'.format(
+                fps, w, h, input, tmp_folder)
+            os.system(cmd)
+
+            # use USR to super resolutioin
+            cmd = 'cd {}/3rdparty/VideoPhotoRepair;python main.py {} --operation usr --scale {} --clc'.format(
+                os.path.dirname(__file__), os.path.abspath(tmp_folder), int(up_rate) )
+            os.system(cmd)
+
+            # convert png to yuv
+            newFileName = self.generateFileName(int(w*up_rate), int(h*up_rate), fps)
+            output = os.path.join(output, newFileName)
+            cmd = 'ffmpeg -i {}/3rdparty/VideoPhotoRepair/results/usr/%d.png -pix_fmt yuv420p {} -hide_banner'.format(
+                os.path.dirname(__file__), output)
+            os.system(cmd)       
+            shutil.rmtree(tmp_folder)
         return output
